+++ core/src/spclient.rs
@@ -8,6 +8,7 @@ use crate::{
     Error, FileId, SpotifyId, SpotifyUri,
     apresolve::SocketAddress,
     config::SessionConfig,
+    dealer::protocol::TransferOptions,
     error::ErrorKind,
     protocol::{
         autoplay_context_request::AutoplayContextRequest,
@@ -901,5 +902,33 @@ impl SpClient {
         );
 
         self.request(&Method::GET, &endpoint, None, None).await
+    }
+
+    /// Triggers the transfers of the playback from one device to another
+    ///
+    /// Using the same `device_id` for `from_device_id` and `to_device_id`, initiates the transfer
+    /// from the currently active device.
+    pub async fn transfer(
+        &self,
+        from_device_id: &str,
+        to_device_id: &str,
+        transfer_options: Option<&TransferOptions>,
+    ) -> SpClientResult {
+        let transfer_options = transfer_options
+            .as_ref()
+            .map(serde_json::to_string)
+            .transpose()?;
+        let body = transfer_options.map(|op| format!(r#"{{ "transfer_options": {op} }}"#));
+
+        let endpoint =
+            format!("/connect-state/v1/connect/transfer/from/{from_device_id}/to/{to_device_id}");
+        self.request_with_options(
+            &Method::POST,
+            &endpoint,
+            None,
+            body.as_deref().map(|s| s.as_bytes()),
+            &NO_METRICS_AND_SALT,
+        )
+        .await
     }
 }